+++ .gitignore
@@ -9,3 +9,6 @@
 
 .vscode
 
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
