+++ requirements.txt
@@ -21,6 +21,7 @@ lark-parser==0.12.0
 llvmlite==0.42.0
 locket==1.0.0
 MarkupSafe==2.1.5
+matplotlib
 mpmath==1.3.0
 networkx==3.3
 numba==0.59.1
@@ -57,7 +58,8 @@ quantum-xir==0.2.2
 requests==2.32.3
 retworkx==0.14.2
 rustworkx==0.14.2
-scikit-learn==1.5.0
+# scikit-learn==1.5.0
+scikit-learn
 scipy==1.13.1
 semantic-version==2.10.0
 six==1.16.0
